+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - `docker` command to build a docker image for local deployment
 - REST API: jwt authentication based on JWT scopes array
+- REST API: pagination for inventory
+- REST API: POST `/repos/export` creates a zip for the specified repository
+- REST API: GET `/repos/export` to download archived repo
+- Added `filter.latest` parameter to REST API `/inventory` listing
 - generating `manufacturers.txt` and `mpns.txt` for improving static hosting (e.g. github, github pages)
 
 ### Fixed